requirements.txt
Flask==1.0.2
python_firebase==1.2
requests==2.19.1
